+++ action/Response.php
@@ -197,31 +197,6 @@ class Response extends \lithium\net\http\Response {
 			$code ? header($header, false, $code) : header($header, false);
 		}
 	}
-
-	/* Deprecated / BC */
-
-	/**
-	 * Expands on `\net\http\Message::headers()` with some magic conversions for shorthand headers.
-	 *
-	 * @deprecated This method will be removed in a future version. Note that the parent `header()`
-	 *             wil continue to exist.
-	 * @param string|array $key
-	 * @param mixed $value
-	 * @param boolean $replace
-	 * @return mixed
-	 */
-	public function headers($key = null, $value = null, $replace = true) {
-		if ($key === 'download' || $key === 'Download') {
-			$message  = "Shorthand header `Download` with `<FILENAME>` has been deprecated ";
-			$message .= "because it's too magic. Please use `Content-Disposition` ";
-			$message .= "with `attachment; filename=\"<FILENAME>\"` instead.";
-			trigger_error($message, E_USER_DEPRECATED);
-
-			$key = 'Content-Disposition';
-			$value = 'attachment; filename="' . $value . '"';
-		}
-		return parent::headers($key, $value, $replace);
-	}
 }
 
 ?>