+++ storage/cache/adapter/Apc.php
@@ -9,18 +9,6 @@
 
 namespace lithium\storage\cache\adapter;
 
-if (!class_exists('\lithium\storage\cache\Adapter')) {
-	$message  = 'The file (`' . __FILE__ . '`) has probably been included directly within ';
-	$message .= 'the bootstrap process. Due to its dependencies it should be (auto-)loaded ';
-	$message .= 'i.e. through the `Libraries` class exclusively. Please update your app\'s ';
-	$message .= 'bootstrap directory to the most recent version or remove the line where ';
-	$message .= 'this file was originally included. ';
-	$message .= 'See https://git.io/voxgh for the most recent version.';
-	trigger_error($message, E_USER_DEPRECATED);
-
-	require_once LITHIUM_LIBRARY_PATH . '/lithium/storage/cache/Adapter.php';
-}
-
 use lithium\storage\Cache;
 
 /**